backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
openai>=1.50.0
pandas>=2.2.0
openpyxl>=3.1.0
xlrd>=2.0.0
python-dotenv>=1.0.0
